VFB-490: refactor directories tree
